indexer: extract module call-related jobs
